move title and exerpt generation (classic) out as a separate plugin
