+++ includes/Classifai/Features/TitleGeneration.php
@@ -231,23 +231,23 @@ class TitleGeneration extends Feature {
 			if ( $screen && ! $screen->is_block_editor() ) {
 				if ( post_type_supports( $screen->post_type, 'title' ) ) {
 					wp_enqueue_style(
-						'classifai-generate-title-classic-css',
-						CLASSIFAI_PLUGIN_URL . 'dist/generate-title-classic.css',
+						'classifai-plugin-classic-title-generation-css',
+						CLASSIFAI_PLUGIN_URL . 'dist/classifai-plugin-classic-title-generation.css',
 						[],
-						get_asset_info( 'generate-title-classic', 'version' ),
+						get_asset_info( 'classifai-plugin-classic-title-generation', 'version' ),
 						'all'
 					);
 
 					wp_enqueue_script(
-						'classifai-generate-title-classic-js',
-						CLASSIFAI_PLUGIN_URL . 'dist/generate-title-classic.js',
-						array_merge( get_asset_info( 'generate-title-classic', 'dependencies' ), array( 'wp-api' ) ),
+						'classifai-plugin-classic-title-generation-js',
+						CLASSIFAI_PLUGIN_URL . 'dist/classifai-plugin-classic-title-generation.js',
+						array_merge( get_asset_info( 'classifai-plugin-classic-title-generation', 'dependencies' ), array( 'wp-api' ) ),
 						get_asset_info( 'generate-title-classic', 'version' ),
 						true
 					);
 
 					wp_add_inline_script(
-						'classifai-generate-title-classic-js',
+						'classifai-plugin-classic-title-generation-js',
 						sprintf(
 							'var classifaiChatGPTData = %s;',
 							wp_json_encode( $this->get_localised_vars() )
